Closes #11: Add update Maven Group repository feature
